Refactor remote project tool executor ownership
