+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -9,7 +9,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `e2e:playwright` | 48 | 7179 | 95 | A | `core.spec.ts` A-<br>`interaction-audit-edge-controls.spec.ts` A-<br>`real-world-actions.spec.ts` A- |
 | `scripts` | 23 | 4008 | 98 | A+ | `native-desktop-smoke.sh` A<br>`smoke.sh` A<br>`live-tr-smoke.sh` A+ |
 | `source:QuillCodeAgent` | 50 | 5952 | 98 | A+ | `AgentToolStepRunner.swift` A<br>`MockPullRequestIntentPlanner.swift` A<br>`AgentBrowserToolAnswerFormatters.swift` A |
-| `source:QuillCodeApp` | 420 | 47165 | 98 | A+ | `WorkspaceRemoteProjectToolExecutor.swift` A<br>`ProjectExtensionManifestLoader.swift` A<br>`QuillCodeToolArtifactViews.swift` A |
+| `source:QuillCodeApp` | 423 | 47182 | 98 | A+ | `ProjectExtensionManifestLoader.swift` A<br>`QuillCodeToolArtifactViews.swift` A<br>`QuillCodeReviewPaneView.swift` A |
 | `source:QuillCodeCore` | 29 | 2222 | 99 | A+ | `ModelInfo.swift` A<br>`AppConfig.swift` A+<br>`AutomationModels.swift` A+ |
 | `source:QuillCodePersistence` | 6 | 354 | 100 | A+ | `ConfigStore.swift` A+<br>`SecretStore.swift` A+<br>`ThreadStore.swift` A+ |
 | `source:QuillCodeSafety` | 7 | 1113 | 99 | A+ | `Safety.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+<br>`StaticSafetyDownloadPolicy.swift` A+ |
@@ -22,7 +22,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `test:QuillCodeAppTests` | 210 | 31466 | 98 | A+ | `WorkspaceRemoteProjectToolExecutorTests.swift` A-<br>`WorkspaceBrowserIntegrationTests.swift` A-<br>`WorkspaceReviewIntegrationTests.swift` A- |
 | `test:QuillCodeCoreTests` | 3 | 524 | 98 | A+ | `CoreModelTests.swift` A+<br>`VerificationResultParserTests.swift` A+<br>`ModelTokenUsageTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 3 | 1108 | 97 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
-| `test:QuillCodeParityTests` | 145 | 13585 | 99 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A-<br>`ParityWorkspaceReviewSurfaceGateTests.swift` A-<br>`ParityMCPGateTests.swift` A- |
+| `test:QuillCodeParityTests` | 145 | 13603 | 99 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A-<br>`ParityWorkspaceReviewSurfaceGateTests.swift` A-<br>`ParityMCPGateTests.swift` A- |
 | `test:QuillCodePersistenceTests` | 6 | 357 | 100 | A+ | `JSONProjectStoreTests.swift` A+<br>`ConfigStoreTests.swift` A+<br>`JSONAutomationStoreTests.swift` A+ |
 | `test:QuillCodeSafetyTests` | 5 | 1211 | 96 | A+ | `SafetyGitPullRequestPolicyTests.swift` A-<br>`SafetyDownloadPolicyTests.swift` A-<br>`SafetyGeneralPolicyTests.swift` A+ |
 | `test:QuillCodeToolsTests` | 26 | 3436 | 99 | A+ | `MCPStdioProberTests.swift` A<br>`ShellToolExecutorTests.swift` A<br>`FileToolExecutorTests.swift` A+ |
@@ -547,12 +547,15 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 99 | 40 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteGitPushCommandBuilder.swift` | 2 lines >120 chars |
 | A+ | 100 | 58 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteGitToolRequestPlanner.swift` | no automated issues |
 | A | 92 | 144 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteGitWorktreeCommandBuilder.swift` | high duplicate-line ratio (26%) |
+| A+ | 100 | 219 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectFileToolExecutor.swift` | no automated issues |
 | A+ | 100 | 29 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectMemoryCommandRunner.swift` | no automated issues |
 | A+ | 100 | 40 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectMemoryTarget.swift` | no automated issues |
 | A+ | 100 | 25 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectMemoryUpdateError.swift` | no automated issues |
 | A | 92 | 140 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectMemoryUpdater.swift` | high duplicate-line ratio (51%) |
 | A | 92 | 105 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectPath.swift` | high duplicate-line ratio (32%) |
-| A | 92 | 436 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectToolExecutor.swift` | watch file size (436 lines); 1 lines >120 chars; duplicate-line ratio (18%) |
+| A+ | 100 | 44 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectToolCatalog.swift` | no automated issues |
+| A+ | 100 | 20 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectToolExecutionContext.swift` | no automated issues |
+| A+ | 96 | 170 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteProjectToolExecutor.swift` | duplicate-line ratio (24%) |
 | A+ | 100 | 9 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRemoteShellCommandFormatter.swift` | no automated issues |
 | A+ | 100 | 23 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceRetryPlanner.swift` | no automated issues |
 | A+ | 100 | 38 | `source:QuillCodeApp` | `Sources/QuillCodeApp/WorkspaceReviewActionRunner.swift` | no automated issues |
@@ -1075,7 +1078,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 36 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightAutomationGateTests.swift` | no automated issues |
 | A+ | 100 | 83 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightCommandPaletteGateTests.swift` | no automated issues |
 | A+ | 100 | 38 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityPlaywrightSettingsRuntimeGateTests.swift` | no automated issues |
-| A+ | 97 | 92 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRemoteProjectToolGateTests.swift` | 16 lines >120 chars |
+| A+ | 97 | 110 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRemoteProjectToolGateTests.swift` | 16 lines >120 chars |
 | A+ | 100 | 34 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRenderedCommandRoutingGateTests.swift` | no automated issues |
 | A+ | 100 | 119 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRenderedCriticalTargetRegistryGateTests.swift` | no automated issues |
 | A+ | 100 | 85 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityRenderedResponsiveTargetGateTests.swift` | no automated issues |
